- code optimization & cleanup - project plan update

+++ common/src/main/java/ru/coolsoft/common/Defaults.java
@@ -2,5 +2,5 @@ package ru.coolsoft.common;
 
 public class Defaults {
 
-    public static final int SERVER_PORT = 7007;
+    public static final short SERVER_PORT = 7007;
 }
